implemented algorithm's pipeline prototype

+++ src/Register.h
@@ -1,4 +1,10 @@
+#include <stdlib.h>
+/**
+ * @brief 
+ *  @param value->Register number
+ *  @param busyBy->Shows which reservation station is using this register. If -1, then it's free.
+ */
 typedef struct {
-    int value;
-    int busyBy;    
+    short value;
+    short busyBy;    
 } Register;
